requirements.txt
# Core dependencies
streamlit>=1.45.0
requests>=2.31.0
python-dotenv>=1.0.0
google-adk 
litellm
clarifai

